Move FluxMaps.read and FluxMaps.write
