+++ gammapy/estimators/flux_map.py
@@ -67,58 +67,3 @@ class FluxMaps(FluxEstimate):
         """Default reference model: a point source with index = 2  (`SkyModel`)"""
         return SkyModel.create("pl", "point")
 
-    def write(
-        self, filename, filename_model=None, overwrite=False, sed_type="likelihood"
-    ):
-        """Write flux map to file.
-
-        Parameters
-        ----------
-        filename : str
-            Filename to write to.
-        filename_model : str
-            Filename of the model (yaml format).
-            If None, keep string before '.' and add '_model.yaml' suffix
-        overwrite : bool
-            Overwrite file if it exists.
-        sed_type : str
-            sed type to convert to. Default is `likelihood`
-        """
-        filename = make_path(filename)
-
-        if filename_model is None:
-            name_string = filename.as_posix()
-            for suffix in filename.suffixes:
-                name_string.replace(suffix, "")
-            filename_model = name_string + "_model.yaml"
-
-        filename_model = make_path(filename_model)
-
-        hdulist = self.to_hdulist(sed_type)
-
-        models = Models(self.reference_model)
-        models.write(filename_model, overwrite=overwrite)
-        hdulist[0].header["MODEL"] = filename_model.as_posix()
-
-        hdulist.writeto(filename, overwrite=overwrite)
-
-    @classmethod
-    def read(cls, filename):
-        """Read map dataset from file.
-
-        Parameters
-        ----------
-        filename : str
-            Filename to read from.
-
-        Returns
-        -------
-        flux_maps : `~gammapy.estimators.FluxMaps`
-            Flux maps object.
-        """
-        with fits.open(str(make_path(filename)), memmap=False) as hdulist:
-            return cls.from_hdulist(hdulist)
-
-    # TODO: should we allow this?
-    def __getitem__(self, item):
-        return getattr(self, item)
